refactor(ucd-store): update import paths for flattenFilePaths
* Changed import paths for `flattenFilePaths` in multiple files to improve module organization.
* Ensures consistency in file structure and enhances maintainability.
