+++ packages/ucd-store/src/internal/analyze.ts
@@ -2,7 +2,7 @@ import type { UCDClient, UnicodeVersionFile } from "@ucdjs/fetch";
 import type { FileSystemBridge } from "@ucdjs/utils/fs-bridge";
 import { promiseRetry } from "@luxass/utils";
 import { ApiResponseError } from "@ucdjs/fetch";
-import { flattenFilePaths } from "./helpers";
+import { flattenFilePaths } from "./flatten";
 
 export interface VersionAnalysis {
   /** Unicode version */
