+++ packages/ucd-store/src/internal/internal.ts
@@ -5,7 +5,7 @@ import type { DownloadError, MirrorOptions } from "../internal/mirror";
 import path, { dirname } from "node:path";
 import { hasUCDFolderPath } from "@luxass/unicode-utils-new";
 import { UCDJS_API_BASE_URL } from "@ucdjs/env";
-import { flattenFilePaths } from "../helpers";
+import { flattenFilePaths } from "./flatten";
 
 type internal__MirrorUnicodeVersionOptions = Required<Omit<MirrorOptions, "versions" | "patterns">> & {
   client: ReturnType<typeof createClient>;
